"10:05am. I am up. In contrast to yesterday, I had some good rest today.
As is usual, the reversible NNs are my main topic of obsession. I have them mostly worked out.

It is amazing that it is possible to come up with a hierarchical algorithm that allows stable bases all the way through the optimization process. I am now more certain than ever that reversibility is one of the main principles of learning. Reversibility itself is credit assignment. It is also what enables modularity in the context of learning. Reversibility is what would allow different kinds of learning algorithms to coexist within the same model.

As great as it is, in its current form it has two very severe flaws.

Namely, that for my theory to work, it needs to be a full batch algorithm. This would make it work on Mnist, but not on anything bigger, and in RL where batches might be potentially infinite that would be a pretty severe restriction.

The second is over fitting. Unless I can make it Bayesian somehow, I am not sure how I am going to be able to do Q learning or any kind of frequentist updates with it. Remember the `1,000,000`, `4,000,000`, `9,000,000` examples? Now suppose I only had `9,000,000` and had to use that as the basis to estimate the rest - that would lead to very poor performance in RL.

RL depends on being iterative and on tricks like optimistic initialization to get exploration. Exact inference that reversibility does would break that.

10:15am. Sigh, it is complicated. Coming this far has just made me realize that I need to solve the rest of the issues as well before I can proceed.

10:20am. I thought I had made an accomplishment, but there is just so much left to go. I am going to have to put more thought into it. But with this, I already beyond what the field understands. There are people looking into reversibility, sure, but at this point in time no one really understand that reversibility is one of the principles.

10:25am. I was going to do a literature review and read some ML papers today, but now that I realize that I am short, I think I will just continue with the lectures. I might as well do this and then learn topology. I think I'll leave the complete elaboration of the learning using reversibility for NNs for the monthly review.

Let me start the lectures."