"10:30am. https://youtu.be/h6bmAhR1fyU?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=706
I am going to rewatch the last one from yesterday. My focus was so low then that I did not get much out of it.

10:50am. https://youtu.be/h6bmAhR1fyU?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1161

He goes a bit fast here. `p(0)` is the first element of the polynumber right? I recognize this as a variant of Taylor's theorem, but I am struggling to see what is going on here.

11:05am. No, reversibility in the frequentist domain makes zero sense. I will never be able to manage it. The updates need to be Bayesian. Forget any other way of doing it. I will forget about LSTD and Zap, and look for Bayesian linear updates of everything.

Reversibility pretty much implies the Bayesian perspective.

I guess Maas was right that the brain both produces noise (inputs acting as targets for the previous layer involves maximizing their entropy) and that it does sampling of network configurations.

I think that even layerwise Bayesian approximation will do as reversibility will modularize the layers.

11:10am. If reversibility is the primary principle of learning, then Bayesity is the secondary.

11:15am. Though I've gotten far better at understanding Bayesian inference, I haven't put nowhere close to enough effort into it as I did for deep learning. In fact I thought it was hogwash for much of my journey.

I am going to have to get deeply familiar with linear Bayesian models.

11:20am. There is just one thing I have to note.

I am really wondering how I am going to optimize the inputs. At first I wrote that I will be able to sample them from the generative model, but no. The lower layers are the generative model. I actually can't sample them.

I'll actually have to optimize them so they meet the targets.

It is interesting that now that I've decided that the intermediate nodes require Bayesian models, that it becomes obvious they also require direct optimization for credit assignment.

What I may have discovered is the point where Bayesian models meet with differentiability. In semi-linear models like I am thinking of, differentiability is required to propagate the targets backward. So backprop would be needed for credit assignment after all as I'll need to adjust the inputs gradually through the non-linearity.

11:30am. If I can find Bayesian updates for linear models that work in an iterative fashion I will be able to make all of this work.

11:35am. I'll unify all the disparate pieces into one coherent design.

11:45am. I do not think there will be anything as convenient as an iterative update for a Bayesian linear layer, but I guess it would not hurt to ask. There should be something.

I am going to let these thoughts simmer for a while and once I elaborate them into a written piece, post them on the ML sub and ask for feedback. This is too good to just leave for the monthly review.

12pm. I've been so caught up in my thoughts that not only did I not watch a single lecture, I haven't even finished rewatching the one from yesterday.

So far September is not at all the way I imagined it would be. I thought it would take me a long time to make anything of the reversibility idea I wrote in the last month's monthly update, but the inspiration just keeping coming and coming. It is positively gushing forth!

12:10pm. Ahhh, I'll never get to the lecture at this rate. I am too in love with these ideas!

Let me have breakfast here."