"1:40pm. Let me resume. I still do not feel like watching the lectures, but I need to start. I keep composing the pieces for that ML post in my head.
2pm. Finally done with yesterday's lecture. I can finally start today's ones. Right now I am really anticipating the points and counter points to my proposal in my head. I do that a lot when I am inspired like now.

2:10pm. Ah, inspiration is so great. It is like a drug.

I really do have so much to learn. Maybe there is hope to this quest after all. Maybe I will be able to do it.

In my desperation, I've made reversibility a principle and it turns out it fits somehow as the basis of credit assignment.

The Singularity or not, I can't go forward by constantly failing. If it is at all possible, I really should be able to master toy tasks like poker with solely the use of my programming ability. If I am to consider myself a good programmer, I should be able to do at least the basic learning tasks.

2:15pm. Things like causing the Singularity feels me with a sense of loneliness when I cannot do even much less than that.

When I threw in the towel at the start of the year and accepted failure, I had no idea what I should do. I had that exploration idea, but I knew it could not possible deal with the stability issues. I just did not want to do yet another dumb experiment that goes nowhere. I got emotionally invested in past ones and was sorely disappointed when they turned out to be nothing.

The worst is that failures essentially seem arbitrary. They are like buggy programs who crash for arbitrary reasons.

With this I might be able to put my failures in the proper context.

When I was a kid I used to do cheats a lot and at time thought how it would be great if I had an agent that would be able to beat the games. At that time, I had absolutely no idea how to program the computers to do such a thing.

2:20pm. It was destiny that I would be here at this age, trying to turn the computer into one of the living. I could never have succeeded at anything else, not with my mindset.

I just mistrust society too much to ever be a part of it. It was set in stone that I would reach the conclusion that it needs to be brought under control if I am ever to be free.

2:25pm. I will conquer this and put the legacy of failure behind me. I might have been a failure for this long, but this long is really a drop in the bucket for somebody who will potentially live for an eternity. One day when I have the power and get challenged on whether I deserve it, I will just produce a window to the person I was now. And point out that I was the insight.

2:35pm. I am having difficulty unchaining myself from the obsession with this idea, but my mind will get tired of it eventually and let me resume the lectures. The poker agent can wait. I can afford to be a failure for a while longer.

Since I am on a math streak, I won't go back to it until I've done all that I wanted to here. I'll do everything properly and by the rules. Just because I have an idea no matter how good, does not mean I can dive into it half cocked. It will get done when its turn comes. It might be a few months down the road as it will take me a while to cover all the ground that I want to first.

2:40pm. Let me see if I can put some headway into the first lecture. I can't just dwell on what is inside my head all day."