"8:10am. I am up. I really am obsessed about the reversibility idea. I keep thinking maybe instead of optimizing the inputs, in a Bayesian setting it might be better to sample them. You can't actually sample the inputs themselves, but you could sample the weights from the lower layer and generate new inputs that way. That is what I keep thinking, but I think optimizing towards the lowest error would still be the right thing to do.
Forget it. Let me do my morning chilling and then I will read the LRA-E paper.

There is something about the reversibility idea.

8:20am. It is fundamentally right.

8:25am. Let me start chilling for real.

9:10am. https://jaiminisbox.com/reader/read/burning-effect/en/0/3/page/1

I like the MC of this, but I'll leave it for later.

Right now let me focus on the LRA-E paper.

I need to take back something I said yesterday, that it would optimize from zero init. My own method actually would not, but I can see why the one in the paper does.

https://wvvw.aaai.org/ojs/index.php/AAAI/article/view/4389
Biologically Motivated Algorithms for Propagating Local Target Representations

Let me read that paper, figure out MC-CFR and then I will get back to the math lectures. I might have lead for the first time towards getting back into RL, but the pieces are hardly all there. I just solved a few of particularly thorny issues, but numerous other still remain. Many details of the algorithm are still need to be elucidated.

9:30am. "Algorithms within the Discrepancy Reduction (Ororbia II et al. 2017) family ..."

Ok, so what I've invented is essentially a DR algorithm. I'll follow up with that paper. Somehow Ororbia is familiar to me. Was he the author of that paper on biological learning that I've read before? No...I've checked his paper list and I do not think he is the guy I am thinking of.

10:55am. Done with the breakfast. I do not like this second paper by the author that much. They are overoptimizing the architecture of it on Mnist. Solve RL and then do what nature did.

I am not a fan of biological plausibility. Nobody really knows what the brain does so these kinds of arguments are not more than imitation of form and not substance.

My own version has synchronization aspects that their paper does not.

10:55am. I said I would check out other papers by him, but forget that for now. Let me read the Pluribus paper and then I will make an update to that paper I wrote.

11:20am. The Pluribus paper has too much stuff in it.

https://arxiv.org/pdf/1809.03057.pdf
Variance Reduction in Monte Carlo Counterfactual Regret Minimization (VR-MCCFR) for Extensive Form Games using Baselines

Let me check this one out.

12pm. "In contrast to RL environments, our experiments in imperfect information games suggest that state-action baselines are superior to state baselines."

This is an important paper. I am going to remember it for future reference and look up code when I need it. I can't really follow the thing as it is complex, and I do not want to get completely in the zone which I'd need to internalize it.

Going into the zone is rather expensive energy-wise for me and I do not want to spend the next few days just messing with this. I pick it up when the time comes along with the rest of this material.

https://mangadex.org/title/26399/tobaku-datenroku-kaiji-24-oku-dasshutsu-hen

Well, since another chapter of Kaiji is out let me take a short break here. I did not get a chance to do so during breakfast.

12:05pm. Ever since I understood the concept of Bayesian inference I've been going back and forth in my head between frequentist and Bayesian reasoning. And conceptually, I am starting to realize that the Bayesian side in contrast to what I thought originally is the one that is easy to reason about.

Bayesians have a well defined notion of convergence, while frequentists have a very hacky notions. On the surface the frequentist methods are computationally tractable, but I am having really hard time making any headway into unearthing their secrets.

I need to go beyond them.

To do that, I need to learn more and give myself a proper mathematical foundation.

Math is what I need right now, not algorithms or programming. Oh yes, let me also update that paper when I get back and post it in a separate file. I'll link to it in the monthly review."