"8:15am. Somehow remembering that it is 2019 sends a ripple of fear through my body. Remembering the old games makes me realize that the future ones will go nowhere unless I push them myself.
I need to keep going.

As fun as what I wrote in that paper was, I really was hoping to get something from topology. Let me start the math lectures.

9:35am. https://youtu.be/8om99RRUBSs?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Let me finally move to number two.

This is not so bad. Today and two more days after that and I will be done with the course and be able to move to topology.

10:35am. Had to take a break. I guess I am not done with thinking about reversibility. I love the idea, but to make it have sense it does need to be Bayesian after all.

Even with a fully reversible activation like PrRelu, necessarily inverting it is not the right move to take. Because the noise from the upper layers, might mean that the inverse results in absurd values. So optimization is the solution, but no matter how I look at it, this optimization is mostly noise. It is all noise.

So the right reverse is Bayesian conditioning.

Except now I would need to do a posterior update of two layers simultaneously, but I am not sure I can even do one.

I had the weight sampling idea. Would doing it then be the right choice. I went back and forth between it yesterday and concluded that optimization would be better, but if one is trying to approximate Bayesian solutions then...

...No, what would be needed is to sample from the updated posterior. That would be what makes sense.

10:45am. I am thinking and thinking and thinking...and I am defeated. I need for things to be more Bayesian for what I want to do to make sense.

10:50am. Discrepancy reduction is the key, but it does need to be Bayesian.

I really hope I could get something from topology to help me out here. I never really payed attention or care for this stuff before now. I really focused way too much on deep learning.

But now I know better. I do not need some trick. I need to master the fundamentals thoroughly.

I will cover all the areas of math from top to bottom and maybe after that some way of doing this will come to me. There is no other way.

Let me focus on calculus. I need to get back to it.

12:40pm. Fuck. I cannot focus. I've been thinking about that idea for the past few hours.

I've realized something important. From the Bayesian perspective, optimizing the inputs makes zero sense.

What does make sense however is updating two layers in tandem. And then going back to frequentist perspective that actually makes the entire reasoning so much clearer.

Both me and the LRA paper are wrong. Two layers should be optimized at the same time. The inputs definitely should not be optimized directly. Rather once the optimization is done, what should be done it to just run the lower layer again to recalculate the input and then use it as a target for next in line.

As much as I dislike backprop, optimizing a series of two layer nets is not beyond reason.

Reversibility does not make sense. There are just so many scenarios where it could go wrong - even in the case that the activation has unbounded range and are reversible like PrRelu and the weights have inverses, those inverses might cause great instabilities due to the noise in the top layer which can never be exactly reversible.

So my conjecture is dead, but the reasoning that lead me in this direction is sound.

Reversibility is not consistent, but optimizing two chained nodes iteratively is. It will get me all the properties of learning that I want.

12:50pm. That having said, from a Bayesian perspective that raises the stakes for updating the posterior even further. I wasn't even sure I could do the usual thing, and now this is coming out.

Ah, what a bother. Maybe I will just end up using factored backprop, plus some learning rate adaptation in the end.

12:55pm. I need to unwind. Let stop here for breakfast."