"1:10pm. Let me resume where I left off. I can always slack more later. Today my goal will be to do 3 more. That will leave the last 5 for tomorrow.
1:20pm. https://youtu.be/Qls3AdgCT3U?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=555

I really need to go to the previous video for a bit. What was the formula for the signed area from the y axis again?

...Ah, right now I remember. The formulas weren't given explicitly so I forgot what things were.

1:50pm. https://youtu.be/Qls3AdgCT3U?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1863

When he says that integration extends only partially, I bet this has a parallel with my reversibility ideas.

1:55pm. https://youtu.be/MgbCV561Qtw?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Two more left in the day.

I want it. I want to get better at math. For the first time in my life I want this seriously enough to put in a long effort.

I need mathematical breakthroughs to do machine learning properly. I need to be the one to do it. But it is just so unlikely that with my mediocre talent in this area I will be able to do what I need. So the most anyone can do in this situation is to do what one can.

It is a compromise imposed on me by reality.

For omnipotence everything. For the wish the price will be paid. Let me move on to the next lecture.

2:35pm. https://youtu.be/MgbCV561Qtw?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1035

It is time to start making winter preparations so let me stop my studies here for a moment.

3:15pm. https://youtu.be/MgbCV561Qtw?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1203

3:35pm. Uf, I ended up missing most of the work somehow and it just ended up being a break. Well, nevermind. Let me resume.

3:45pm. https://youtu.be/MgbCV561Qtw?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=1661

I am actually really surprised to how much care he puts into talking about undefinedness. And how much relevance this has for doing credit assignment. I was right to dive deeper into math and Wildberger's lectures are well worth listening to.

3:55pm. Right now I am downloading Yukikaze 200mb piece at a time and am asked to solve a capcha. 'Skynet is watching'. That got a laugh out of me.

4:05pm. https://youtu.be/wb1muCVg_YM?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=315

Yeah, I need to understand this. These rigid definitions is what I need. Saying that i^2 = -1 does not make sense, but identifying it with some 2x2 matrix makes everything that much more reasonable.

4:25pm. Damn, I need a break. Right now I am at the 10:30 mark, he is saying something and I can't muster the resources to really think about it. I really do not remember math being this difficult back in school. This is taking quite a bit of my mental stockpile to pay off.

5:35pm. Done with lunch. Let me finally finish the video.

https://youtu.be/wb1muCVg_YM?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB&t=756

Group structure...first time hearing this. But properties such as that would be great to have in mind when doing optimization. I need to go beyond my present knowledge.

Whatever the final thing turns out to be, you'd expect it would be more than just basic calculus and linear algebra. Learning is everything and it is destiny that all the fields of mathematics play a role in its computational development.

6:10pm. https://youtu.be/Mo-y5k3voPU?list=PLzdiPTrEWyz4rKFN541wFKvKPSg5Ea6XB

Ouf, let me stop here for the day. It is amazing how difficult it is to just go through 5 lectures per day. It is really killing me.

6:20pm. Time for some rest."