"12:10pm. Ah, it is finally time to do the thing. Let me get to it.
1:30pm. Done with chores for the day.

2:05pm. Let me resume.

...Though I should be studying topology, I kind of want to see what he says in the next two videos. I'll put some time into topology after that.

2:25pm. https://youtu.be/YYnYatWV-pU?t=928

I forgot all this stuff about repeating decimals.

3pm. https://youtu.be/YYnYatWV-pU?t=2583

Ok, I am just about ready to accept that `1 = 0.9~` after all. But now that I've had this talk I see repeating decimals as a definite step back from fractions. At most what I've been convinced is that they are equivalent to fractions.

3:15pm. That video was long. Let me watch topology for a bit.

There is something that I need to understand. Just what does the standard weight update for a linear layer average exactly? I definitely need to look into least squares.

I admit that I thought random feedback weights were completely nonsensical, but maybe there is some theory there. Without the weights constantly shifting, it might be easier to come to a conclusion as to what credit assignment is doing.

Let me just note something. Even with optimizing `W x = y` using SGD, adding KFAC to reproject the gradients...assuming those inverse covariances are static does not change the theory of figuring out what the update does. At most what higher order optimizers should be doing is speeding up the time it takes a first order optimizer to arrive at the same answer. Fundamentally, they should not be different answers.

4:35pm. As is usual, I am caught up in my own thinking rather than watching the topology course.

4:55pm. Forget about entropy. Just one video and I am done for the day. I keep thinking of what the backprop reverse updates do and I think I am getting a bit of feel for it.

5:30pm. Ah forget it. I can't focus a bit. When I am watching the video I just seem to be thinking about that problem and in my laziness even gazing at manga. I clearly do not feel like doing any more work, if you can call this work.

Well, it is work for me.

I am starting to sense that I am going to have to get to the bottom of frequentist updates, the optimization difficulties be damned.

5:40pm. I feel that there is an intimate connection between entropy maximization and nonlinearities. All the higher order optimization tricks that I know just point to this one conclusion.

Right now I've realized that in a RL setting, when the only thing you have is one continuous value as a target (as opposed in SL when you usually have a vector label) when combined with backprop updates would lead to entropy collapse.

The reason why that is, is because in the last layer the kind of gradients resulting from the weight update would just be the weight values rescaled.

5:50pm. Mhhhh...

How should whitening of inputs be done in the framework of discrepancy reduction?

How would one both whiten inputs and optimize with the respect to the layer weights?

6:05pm. I really need to pry myself off. Forget those pressing questions and instead play and rest for a while."