"3:45pm. I am back. Let me read Kengan Omega and then I will watch a few math fundamental lectures. Tomorrow I will stop messing about and resume going through PRML. I'll go from start to the end by the end of the month or earlier. I won't waste too much time going through the exercises.
4:10pm. No, no. I can't stomach any more lectures. I want to read the PRML book period. I had enough of lectures. Let me finally start what is left of the day.

4:45pm. Done with lunch. I am now looking at the solution for exercise 1.7. Months ago, I spent a lot of time studying this.

5pm. Wow, it was not just me. This is so incredibly convoluted.

5:35pm. 48/748. I remember why back in 2015 I never got far into this book. If I did not study probabilistic programming, I would find this stuff incredibly convoluted. The notation of the Bayesian field is definitely amongst the worst of all the fields of mathematics. It is so bad that it is almost meaningless.

6:05pm. 55/748. The last time I read this I remember being confused as to what the radius of a Gaussian means, but it probably means the distance from the mean.

6:10pm. Mhhh...I realized I've been skipping the exercises, but the book's content so far is not too tough. It would have been impossible as a starting point, but now that I actually know how Bayesian inference is used in terms of probabilistic programming, I can make sense of this.

Forget the dumb math. I'll leave that for later after I am done with the chapter.

I'll aim to finish reading it tomorrow and then go through the exercises. Unlike a few months ago, I understand that this math involving infinite integrals is particularly difficult, so I won't strain myself too much over it.

I want to get the easy parts of this book - and the actual book does not seem to hard and then move to programming. I've paid my dues to the math gods and got the principle of learning as reversibility via optimization.

For the rest of the day it is time to rest. I think I will be able to take it a bit easier starting from here. As little as I did of the lectures, I found going through them highly strenuous. But for my efforts I did gain some skills I did not have before. And soon I will get back to what my strengths are even if I will be using, ugh, Python.

I look forward to it."