"9:45am. I want to watch Machikado Mazoku, but I will leave that for later.
What I will do now is read the PRML book and finish the first chapter. For the next week that is all I want to do.

Just now I had a good idea. Rather than just whitening the targets, I can multiply the weight matrix on the lower layer by the whitening matrix. That might make the optimization problem easier.

Or actually what I could do is reverse the order a bit. First update the targets, train the lower layers ... and then do the whitening after the synchronization.

What this would allow me to do is to directly whiten from inputs the lower layer by multiplying the weights for it with the whitening matrix. That would be sick.

9:50am. Yeah, without a doubt, that is the right thing to do. It is better than my previous idea, but remember that thing about the Cholesky factorization inserting random rotations and such into the weight matrix? I'd have to make sure the synchronization is done properly.

Alternatively, why don't I forget that nonsense and just use KFAC since I want whitening so much? Why struggle over this?

If I am going to go to so many lengths to make whitening work, I might as well just do that.

9:55am. Let me start with the book.

10:40am. Holy shit, how did I spend so much time on the update?

10:50am. I am finally reading it. Oh yeah, yesterday I finally saw the formula for Bayesian regression in the book. I think that was the Gaussian process formula actually. No way to sample the weights from that.

Getting backprop to work will probably be my best bet after all.

Using KFAC to do whitening rather than having to deal with the synchronization issues of explicit variant would be good as well. I thought that the later might have less issues, but it does feel like I am wrong.

I should just do the easier thing first. I can make this thing work.

11:25am. My focus seems to be low again. This book just breezes through a couple of fields worth of material without really explaining anything properly.

12pm. 74/748. I am not sure what to think at this point.

I did so much studying of theorem proving, and studied math in general, but I do not think any of that will help me to figure out what is going on in this book. In fact, I get the sense I was better months ago when I first read the chapter and was really focusing on it. Right now I am just skimming it.

One thing I realize now that I do not from months ago is that me not understanding this is not necessarily an indicator of lack of skill or ability. It might just be the fact that the math itself lacks credibility. There are all sorts of complicated fields that instead of calculate, just waffle about things that can never be calculated.

The intro chapter rather than really teaching me, just touches a few points and then moves on as if I understand everything.

The book itself is trash!

On one end there is theorem proving in Coq and Agda, and lectures like Wildberger's that do everything from the ground up and you can grok it if you understand basic arithmetic.

On the other end there is this. It is like diving into some undocumented large codebase and trying to understand the underlying assumptions just by reading it. I can't do it.

12:20pm. Ah, this is not going as planned. Not at all.

I've really been wondering how exactly learning theorem proving is going to make me better at this kind of math, and as it turns out, it won't. I've been afraid of this from the start.

I admit it. I tend to do wild and crazy things in order to try and open up a path to the next thing, but my pessimistic estimates turn out right way too often.

Still, if I just did things that are right, I'd never move from the spot I am rooted in.

Often it happens to be the case that by exploring you do not find what you seek, but something more valuable. With Cubical Agda as a channel, I think I am going to follow up my view that reversibility is the principle of learning.

12:25pm. I am thinking about those 12 Gaussian proofs of normalization, and they were really incredibly creative and complicated. They are not at all like the Coq and Agda proofs.

But really, the formal logic proofs that I did in the theorem provers are what the golden standard is. They are what is really important. Not the math that is equivalent how many angels can be fit on the head of a pin.

12:30pm. Let me stop here for a bit. I'll have breakfast and then move on to chapter two.

Ultimately I cannot compromise on my principles. Even if it means having to accept that I will have to be dumb, I am going to have to stick to the kind of mathematics that can be verified. Unless I do this, I will get lost fairly quickly. I learned theorem proving and it is not at all pushing me in the direction of the kind of math that can be found in PRML."