"1:55pm. Let me resume. I want to get back into the book rather than just wasting time.
2:05pm. The exercise are far too difficult. Even if I somehow got them, there is no point in it unless I get feedback.

2:45pm. 105/748. This book does not let up in terms of difficulty.

3pm. No, I had enough. My plan too get better at Bayesian inference by getting better at logic is a bust. This book is literally a useless survey and the whole Bayesian field is living on a cloud to make it worse. Wildberger mentioned this field being problematic and now I agree. The only thing I have for my logic efforts are tie ins into programming.

3:05pm. I can't muster any more focus at all. I had enough of lectures and books and papers.

Enough. Enough. Enough!!!

I am done with this. What is done is done. I gave it all a chance.

I'll take what math I can that is logically grounded and dump the rest. Skills should be built from the ground up. It might be different for 'real' mathematicians, but as a programmer, I need the proofs to make sense from the very root.

3:15pm. As much as I hated parts of theorem proving, if that sort of thing was actually useful it would be the ideal of learning. You have a problem and then you try to solve and get feedback every step of the way.

What did I even do for the past month? I can't remember.

Looking at the entry for 8/15, I was starting to study HITs at that time...

3:20pm. Ok, let me take a little breather.

For the past 20 days, ever since I wrote that reversibility rants, I had an image in my head and it has absolutely assaulted me with inspiration. I've managed to solve so many issues related to credit assignment that it is not even funny.

It is time I test those ideas.

It is time for me to return to programming.

It has really been too long. And besides those ideas, I've wanting to look into tabular Bayesian RL and more CFR for a while now. Originally, I wanted to improve my theorem proving skills so I could prove convergence properties of CFR, but I am not sure I can really do that at my current level. So I am going to have to return to my old empirical tricks.

3:25pm. What I really want to do is start with tabular Bayesian RL. The new credit assignment can wait, and I am not too interested in diving into this right away. I am still thinking about it and every day is giving me new insights.

Tabular Bayesian RL might be something cute and practical in certain cases and I am not sure how it could be done at all.

It has been on my mind ever since I read Osband's thesis and I want to retrace that work just to see where it leads.

It should be a good exercise to get me warmed up to programming again.

With CFR, I really should implement the newest Monte Carlo variant and test them on those toy games.

3:30pm. So here is the plan:

1) Tabular Bayesian RL.
2) Monte Carlo CFR.
3) Discrepancy reduction.

On the academic side:

1) Wildberger's lectures.
2) The Cubical library.

3:35pm. Now, I do not feel like starting this right now. It is not too late, but it is kind of late.

Let me take a look at whether there is any code for that first thing.

3:35pm. http://proceedings.mlr.press/v87/morere18a/morere18a.pdf
Bayesian RL for Goal-Only Rewards

"We now seek to obtain a model-free RL algorithm able to explore with few environment interactions, and providing a full predictive distribution on state-action values. Kernel-based method GPTD [20] and Least-Squares TD [27] are among the most data-efficient model-free techniques. While the former suffers from prohibitive computation requirements, the latter offers an appealing trade-off between data-efficiency and complexity. We now derive a Bayesian RL algorithm that combines the strengths of both kernel methods and LSTD."

I am not finding anything on tabular Bayesian RL - I think I might try asking on the RL sub, but this paper came up on the top of Google. This seems interesting.

4pm. https://www.aaai.org/Papers/AAAI/1998/AAAI98-108.pdf
Bayesian Q-learning

I think this was where the PSRL was first introduced.

4:10pm. https://www.reddit.com/r/reinforcementlearning/comments/d6vhdx/does_anyone_have_implementation_examples_of/

I made the post.

4:15pm. So at this point I seem to be reading Osband's thesis again.

https://ceit.aut.ac.ir/~shiry/lecture/machine-learning/papers/BRL-2000.pdf
A bayesian framework for reinforcement learning

Ah, it was not Bayesian Q learning. It was this one.

4:30pm. I have quite a lot of motivation for Osband's thesis. I guess I'll spend what is left of the day refreshing my memory of it. Then I will go through the papers I've linked to.

4:55pm. 129/179. I am confused. The paper talks about the tabular algorithm, but I do not think there is any such thing.

Also, the algorithm on this page seems to have a different network for every timestep.

6:10pm. I got some stuff linked by gwern from that thread I posted.

https://github.com/aslanides/aixijs
https://github.com/mcastron/BBRL

I'll look into it tomorrow. I've gone through the relevant sections for Osband's thesis and the two old papers. They do not have much in terms usable stuff, and they are pretty convoluted as Bayesian ML papers are opt to be.

6:15pm. Still, I do want to follow up this thread to see where it goes.

Now that I am studying this I feel much more motivated. I do not think I'll get far with this thread, but I am decently confident that the thing that comes after this - MC-CFR I should not have too much trouble putting into action.

I have to refresh my memory of how LSTD and Zap work after that, but all in good time.

Also I might want to study the SAC by OpenAI. It might be good to replace the standard AC formulation with something more appropriate. I'll see how that goes as well.

6:20pm. I have a decent bit to go through before I can resolve to start working on the reversibility ideas, but it should be a neat little excursion. Then from October onwards, I will finally sit down and hopefully bring some result by the time of the next update.

For today I am done."