"9:40am. Got up a bit ago. Let me chill for a while and then I will start.
10:05am. Let me start. What should I do first?

https://github.com/aslanides/aixijs
https://github.com/mcastron/BBRL

I'll leave these for later. Agh, somehow I really do not feel like diving into these libraries, but I am going to do it. I'll probably learn something from it.

It will take some messing around to set up C++ and JS environments properly no doubt.

http://proceedings.mlr.press/v87/morere18a/morere18a.pdf
Bayesian RL for Goal-Only Rewards

Let me read this. This seems to be along the lines what I've been looking for.

10:25am. Uah, let me do the chores.

10:50am. "where the variance is integrated over all possible actions, and Vmax = arg maxa V[Q(s0, a)] ensures r^exp is always negative. By combining an optimistic model for U to negative rewards, optimism in the face of uncertainty guarantees efficient exploration."

Hmmm, this choice of making the exploration reward always negative is interesting.

10:55am. I definitely am interested in how could Bayesian linear regression be used for this instead of LSTD.

11:05am. Interesting that they got Sherman-Morrison to work.

11:30am. I do not understand all the aspects of the algorithm presented in the paper. Like the way targets are kept track of. I do not understand why weights are never used, only feature means instead. Nor do I understand the way uncertainty is propagated.

But it is impressive nonetheless. It has all the features I've been thinking off, in one fast package. This is also the first time I've seen Bayesian linear regression used in an online fashion.

I'll definitely be keeping this paper in mind. If it is possible to stably update the layers in an online fashion, rather than just having Bayesian linear regression on the top, it might be usable in the hidden layers. This would make weight sampling viable.

11:40am. I'll dive deeper into the paper later. Forget it for now. Everything in its own time.

https://www.fransoliehoek.net/publications/htmlfiles/b2hd-Katt19AAMAS.html
Bayesian Reinforcement Learning in Factored POMDPs

11:45am. Let me take a look at this and I will take a break.

For the remainder of the day, I will take a look at the libraries starting with AIXIjs.

12:05pm. Actually...Bayesian linear regression requires regularization. How did they manage that?

12:10pm. Hmmm...I can't find any mention of this.

12:15pm. Ok, I am done with `Bayesian Reinforcement Learning in Factored POMDPs`. It is a fairly tough paper and does not have any easy to use algorithms in it. I am not too interested, but maybe I'll look at the code later.

With this let me take a break. The next thing on my schedule will definitely be those two libraries. I guess I will be playing with them for the next few days or so. After that comes CFR.

And after that comes discrepancy reduction in Python + PyTorch. First I will get it to work on the CPU, and once I have the right algorithms, only then will I consider resuming work on Spiral. I have no doubt that neurochips + algorithms will reach a steady state and that is the time to make optimized implementations of it."